Narrow wgs coverage_cap to u16.
The per-position depth array stores u16 counts, so any --coverage-cap
above 65535 never reflects real saturation behavior. Typing the CLI
option as u16 lets clap reject out-of-range values at parse time with a
clear error rather than silently clamping deep inside ContigDepth::new,
and drops the defensive min/try_from dance from the finalize loop.

+++ tests/test_wgs.rs
@@ -20,7 +20,7 @@ fn make_cmd(
     exclude_unpaired_reads: bool,
     min_mapq: u8,
     min_bq: u8,
-    coverage_cap: u32,
+    coverage_cap: u16,
 ) -> Wgs {
     Wgs {
         input: InputOptions { input: bam.to_path_buf() },
